Revamp setup.sh
This revamp allows the script to insert content into files like ~/.vimrc directly from the DUMP repository, ensuring that the inserted configurations are always up to date with the latest version in the repo. It also makes sure that the user has to understand how/when to install the script.